move module data to its own file
