+++ ui/src/App.js
@@ -4,6 +4,7 @@ import './App.css'; // This now imports Tailwind CSS
 import axios from 'axios'; // Import axios for API calls
 import { JsonView, darkStyles } from 'react-json-view-lite';
 import 'react-json-view-lite/dist/index.css';
+import moduleData from './moduleData'; // Import the module data
 
 function App() {
   const [currentStep, setCurrentStep] = useState(0);
@@ -21,88 +22,6 @@ function App() {
 
   useEffect(() => {
     const fetchModules = async () => {
-      // Simulate fetching module data from an API or file
-      const moduleData = [
-        {
-          order: 0,
-          name: "Main Keyword",
-          component: 'MainKeyword',
-          prerequisites: [],
-          executionTime: 0,
-          hasProgressBar: false,
-          buttonLabel: "Get Main Keyword",
-          renderProgressMessage: null
-        },
-        {
-          order: 1,
-          name: "Keyword Research",
-          component: 'KeywordResearch',
-          prerequisites: ["MainKeyword"],
-          executionTime: 60, // in seconds
-          hasProgressBar: true,
-          buttonLabel: "Perform Keyword Research",
-          renderProgressMessage: {
-            5: "Analyzing Keywords...",
-            10: "Generating Keyword Variations...",
-            30: "Using AI to Categorize and Rank Keyword Variations...",
-            60: "Removing Duplicates and Contextually Similar Keyword Variations...",
-            80: "Semantically Grouping Keyword Variations...",
-            99: "Wrapping up...",
-            100: "Complete! ...But it looks like there are a few extra things to iron out, please give it a few more seconds"
-          }
-        },
-        {
-          order: 2,
-          name: "Research Similar Articles",
-          component: 'SimilarArticles',
-          prerequisites: ["KeywordResearch"],
-          executionTime: 90, // in seconds
-          hasProgressBar: true,
-          buttonLabel: "Analyze Articles",
-          renderProgressMessage: null
-        },
-        {
-          order: 3,
-          name: "Select Headings",
-          component: 'Headings',
-          prerequisites: ["SimilarArticles"],
-          executionTime: 45, // in seconds
-          hasProgressBar: true,
-          buttonLabel: "Select Headings",
-          renderProgressMessage: null
-        },
-        {
-          order: 4,
-          name: "Titles",
-          component: 'Titles',
-          prerequisites: ["Headings"],
-          executionTime: 180, // in seconds
-          hasProgressBar: true,
-          buttonLabel: "Select a Title",
-          renderProgressMessage: null
-        },
-        {
-          order: 5,
-          name: "Article Outline",
-          component: 'ArticleOutline',
-          prerequisites: ["Titles"],
-          executionTime: 120, // in seconds
-          hasProgressBar: true,
-          buttonLabel: "Generate Article Outline",
-          renderProgressMessage: null
-        },
-        {
-          order: 6,
-          name: "Article Generation",
-          component: 'ArticleGeneration',
-          prerequisites: ["ArticleOutline"],
-          executionTime: 60, // in seconds
-          hasProgressBar: true,
-          buttonLabel: "Generate Article",
-          renderProgressMessage: null
-        }
-      ];
-
       // Dynamically import the components
       const importedModules = await Promise.all(
         moduleData.map(async (module) => {
